+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 #### Improvements
 * Fix #1874: Added unit tests verifying windows line-ends (CRLF) work
+* Fix #1177: Added support for OpenID Connect token in kubeconfig
 
 #### Dependency Upgrade
 * Update Jackson Bom to 2.10.2